api/requirements.txt
fastapi
uvicorn
python-dotenv
aiohttp
sqlalchemy
psycopg2-binary
asyncpg